Create What is library app?
This is one of the first my projects after a long time. I learn some more about Git, and GitHub and from this moment I will push every my project on GitHub. 
This is a Library aplication that helps people write books that they readed. To make them own library with books to see much did they read and to dont forget...